First commit of coronacrypto project
These scripts are used to generate puzzles for the coronacrypto
project. The main script is ccpuzzler.py. It is used at the CLI
to generate and log a puzzle given plaintext and key generation
parameters. The cipher package includes the modules monoalphabets
and monosequences which manage monoalphabetic keys and sequences
respectively. Their sum total is responsible for key management.
The modules encipher and decipher are responsible for the message
encipherment and decipherment. The later is only used for testing
purposes.

This code was committed along with the release of Series 1, Puzzle 1.
The whole project started as something to do (and something for others
to do) during the Coronavirus Pandemic of 2020.