Added CLI arguments for block/row size and pad
The CLI now accepts arguments for the size of letter blocks,
the number of blocks per row, and the letter to be used
as a pad in ciphertext. These parameters were previously
optional parameters in encipher and defaults were set there.
They are now required positional arguments and defaults are
drawn from the CLI argument parser.